requirements.txt
flask
flask_sqlalchemy
flask-limiter
python-dotenv
pyserial
qrcode[pil]
requests
gunicorn
